requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7.1
Sphinx==1.2.1

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.2.1